Move user api functions into userRouter
