+++ server/routes/user/userRouter.js
@@ -1,0 +1,60 @@
+const userRouter = require('express').Router();
+const { findUserModel, insertUserModel } = require('./userQueries');
+const { validatePostUser } = require('./userValidations');
+
+function responder(res, code, message) {
+  // const functionName = 'responder';
+  res.statusCode = code;
+  res.json(message);
+}
+
+async function processPostUser(body, res) {
+  const functionName = 'processPostUser';
+  try {
+    // debug(`${functionName}, "body",  ${inspect(body)} ${functionName}`);
+    const {
+      // eslint-disable-next-line camelcase
+      email_verified,
+      family_name,
+      given_name,
+      locale,
+      sub,
+      updated_at,
+      ...subSetBody
+    } = body;
+    const keys = Object.keys(subSetBody);
+
+    const findUserResults = await findUserModel(body);
+    const { rowCount } = JSON.parse(JSON.stringify(findUserResults));
+    // debug(`${functionName} findUserResults${inspect(findUserResults)} rowCount ${rowCount}`);
+    if (rowCount === 0) {
+      const insertUserResults = await insertUserModel(subSetBody, keys);
+      // debug("insertUserResults ", await insertUserResults);
+      if (!insertUserResults) {
+        responder(res, 500, { error: 'error saving' });
+        return;
+      }
+      responder(res, 200, { data: await insertUserResults[0] });
+      return;
+    }
+    responder(res, 200, { data: subSetBody });
+    return;
+  } catch (err) {
+    console.error(`CATCH ${functionName}}`);
+    responder(res, 500, { error: err });
+  }
+}
+
+function postUser(req, res) {
+  // const functionName = 'postUser';
+  const validated = validatePostUser(req);
+  if (!validated) {
+    return responder(res, 400, { error: 'not a valid request' });
+  }
+  // debug(`req.body ${inspect(req.body)} ${functionName} HERE`);
+  return processPostUser(req.body, res);
+}
+
+userRouter.post('/', postUser);
+
+module.exports = userRouter;
